Patient UI- configured the messaging system *NEW*
*IMPORTANT* Patient now has an ArrayList of messages that is written
like this: "MM/DD/YYYY <message>". I added some examples in the
PatientUI class. It functions correctly. This is supposed to coordinate
with Matt's Doctor UI portion where a doctor contacts a patient and
types a message. It will be easy on Matt's part because when a doctor
types in the message in the textField (I recommend you use a JTextArea
instead with setEditable(true) ), just get the text from that field/area
and say patient.addMessage(getDate() + " " + textarea.getText());